Merged main into this branch. Fixing conflicts and making changes as needed for the node system to work.

- Made the circuit elements menu dragable.

- Node system now works with as many output nodes as needed.

- Fixed edge case with line drawing on no input slot found. The line manager draw flag wasn't reset.

Everything seems to be working with the node system after merge. Some minor changes are still needed but will push this to main for keeping things up to date.

+++ node_classes/circuitnode.cpp
@@ -1,7 +1,7 @@
 #include "circuitnode.h"
 
 CircuitNode::CircuitNode(int inputCount, bool hasOutput, QWidget *parent)
-{    
+{
     this->setParent(parent);
     this->setObjectName("circuitNode");
     this->setAccessibleName("circuitNode");
@@ -38,7 +38,7 @@ void CircuitNode::moveWidget()
     DragableWidget::moveWidget();
 
     // Update inputs line
-    for (int i = 0; i < inputs.size(); i++) {
+    for (unsigned long i = 0; i < inputs.size(); i++) {
         if (inputs[i]->connection == nullptr) {
             continue;
         }
